+++ examples/libsql-sparql-scale/main.ts
@@ -1,8 +1,7 @@
 import { createClient } from "@libsql/client";
 import { QueryEngine } from "@comunica/query-sparql-rdfjs-lite";
-import { Client } from "@worlds/client";
 import { createComunicaLibsqlSparqlEngineFactory } from "@worlds/client/adapters/comunica";
-import { createLibsqlAdapter } from "@worlds/client/adapters/libsql";
+import { createLibsqlClient } from "@worlds/client/adapters/libsql";
 import { DataFactory } from "n3";
 
 const { quad, namedNode, literal } = DataFactory;
@@ -18,14 +17,12 @@ if (import.meta.main) {
   const databaseClient = createClient({ url: ":memory:" });
   const queryEngine = new QueryEngine();
 
-  const client = new Client(
-    await createLibsqlAdapter({
-      client: databaseClient,
-      createSparqlEngine: createComunicaLibsqlSparqlEngineFactory({
-        queryEngine,
-      }),
+  const client = await createLibsqlClient({
+    client: databaseClient,
+    createSparqlEngine: createComunicaLibsqlSparqlEngineFactory({
+      queryEngine,
     }),
-  );
+  });
 
   await client.import({
     source: {